Chore: clean up unused AST utils
