+++ lib/rules/no-unused-vars.js
@@ -6,34 +6,9 @@
 
 const ruleComposer = require("eslint-rule-composer");
 const utils = require("../utils");
+const { hasSideEffect } = require("../ast-utils");
 
 const rule = utils.getFixableRule("no-unused-vars", false);
-
-/**
- * Check if an expression has side effect.
- *
- * @param {Node} node AST node
- * @returns {boolean} result
- */
-function hasSideEffect(node) {
-    if (["Literal", "Identifier", "ThisExpression"].includes(node.type)) {
-        return false;
-    }
-
-    if (node.type === "MemberExpression") {
-        return hasSideEffect(node.object) || hasSideEffect(node.property);
-    }
-
-    if (node.type === "TemplateLiteral") {
-        return node.expressions.length !== 0;
-    }
-
-    if (node.type === "AssignmentExpression") {
-        return hasSideEffect(node.right);
-    }
-
-    return true;
-}
 
 module.exports = ruleComposer.mapReports(
     rule,
